DW uses PyroCommon API for functions now
