+++ DeadlyWeapons/DeadlyWeapons.csproj
@@ -53,7 +53,6 @@
     </ItemGroup>
     <ItemGroup>
         <Compile Include="DFunctions\ConsoleCommands.cs" />
-        <Compile Include="DFunctions\SimpleFunctions.cs" />
         <Compile Include="DFunctions\VersionChecker.cs" />
         <Compile Include="DFunctions\Wrapper.cs" />
         <Compile Include="Main.cs" />
@@ -70,6 +69,12 @@
     <ItemGroup>
       <None Include="App.config" />
       <None Include="packages.config" />
+    </ItemGroup>
+    <ItemGroup>
+      <ProjectReference Include="..\PyroCommon\PyroCommon.csproj">
+        <Project>{6e27f96f-aa33-477f-80f4-4245324be25c}</Project>
+        <Name>PyroCommon</Name>
+      </ProjectReference>
     </ItemGroup>
     <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
     <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
